# api_server/Stockadoodle-api-docs.yaml
openapi: 3.0.0
info:
  title: StockaDoodle Inventory Management API
  description: >
    RESTful API for the StockaDoodle desktop inventory management system. This
    documentation describes the local API instance running at each individual
    QuickMart store (e.g., Quezon City, Makati, Taguig). 

    Each instance uses a local SQLite database and operates independently.
  version: v1
servers:
  - url: 'http://127.0.0.1:5000/api/v1'
    description: Local Flask API Server for the current store location
paths:
  /products:
    get:
      tags:
        - Products
      summary: Retrieve all products
      description: >-
        Retrieve all products with optional filters (category, search term,
        stock levels).
      parameters:
        - in: query
          name: category_id
          schema:
            type: integer
          description: Filter products by category ID.
        - in: query
          name: search
          schema:
            type: string
          description: Filter by product name or brand.
        - in: query
          name: low_stock
          schema:
            type: boolean
          description: 'If true, filters for products below minimum stock level.'
      responses:
        '200':
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Create a new product
      description: Create new product with required fields including image_blob (base64).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  '/products/{product_id}':
    get:
      tags:
        - Products
      summary: Get single product details
      description: Get single product details including image_blob (base64 encoded).
      parameters:
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
          description: The ID of the product.
      responses:
        '200':
          description: Product details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      tags:
        - Products
      summary: Update all product fields (full replacement)
      description: Full replacement update of all product fields.
      parameters:
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product successfully updated.
    patch:
      tags:
        - Products
      summary: Partial update of product fields
      description: Apply a partial update to one or more product fields.
      parameters:
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Product successfully patched.
    delete:
      tags:
        - Products
      summary: Delete product
      description: Deletes (soft or hard delete) a product record.
      parameters:
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: Product successfully deleted.
...
  '/log/product/{product_id}':
    get:
      tags:
        - Product Logs
      summary: Get audit history for specific product
      parameters:
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: A list of product log entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductLogEntry'
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        category_id:
          type: integer
        brand:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        image_blob:
          type: string
          nullable: true
          description: Base64 encoded image blob for the product.
        expiration_date:
          type: string
          format: date
          nullable: true
        min_stock_level:
          type: integer
      required:
        - name
        - category_id
        - price
        - stock
    ProductInput:
      type: object
      properties:
        name:
          type: string
        category_id:
          type: integer
        brand:
          type: string
        price:
          type: number
          format: float
        stock:
          type: integer
        image_blob:
          type: string
          nullable: true
          description: Base64 encoded image blob for the product.
        expiration_date:
          type: string
          format: date
          nullable: true
        min_stock_level:
          type: integer
      required:
        - name
        - category_id
        - price
        - stock
        - min_stock_level
...
    ProductLogEntry:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        user_id:
          type: integer
        action_type:
          type: string
          enum:
            - SALE
            - DISPOSAL
            - STOCK_ADJUSTMENT
            - CREATION
        details:
          type: object
          description: 'JSON details about the action (e.g., quantity, reason).'
        timestamp:
          type: string
          format: date-time
...